# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HyperMind AI API — multi-agent AI SaaS platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /memory:
    get:
      operationId: listMemories
      tags: [memory]
      summary: List all memory entries for the current user
      responses:
        "200":
          description: Memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"
    post:
      operationId: createMemory
      tags: [memory]
      summary: Store a new memory entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemoryInput"
      responses:
        "201":
          description: Memory created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemoryEntry"
  /memory/{id}:
    delete:
      operationId: deleteMemory
      tags: [memory]
      summary: Delete a memory entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /memory/search:
    get:
      operationId: searchMemory
      tags: [memory]
      summary: Search memory entries by query
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Matching memory entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemoryEntry"

  # ─── Templates ────────────────────────────────────────────────────────────────
  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all prompt templates
      parameters:
        - name: agentId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromptTemplate"
    post:
      operationId: createTemplate
      tags: [templates]
      summary: Create a new prompt template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromptTemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptTemplate"
  /templates/{id}:
    get:
      operationId: getTemplate
      tags: [templates]
      summary: Get a specific prompt template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptTemplate"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteTemplate
      tags: [templates]
      summary: Delete a prompt template
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Files ────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    MemoryEntry:
      type: object
      properties:
        id:
          type: integer
        key:
          type: string
        value:
          type: string
        agentId:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - key
        - value
        - createdAt
    MemoryInput:
      type: object
      properties:
        key:
          type: string
        value:
          type: string
        agentId:
          type: string
      required:
        - key
        - value

    # ─── Templates ───────────────────────────────────────────────────────────
    PromptTemplate:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        content:
          type: string
        agentId:
          type: ["string", "null"]
        category:
          type: string
        isBuiltin:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - content
        - category
        - isBuiltin
        - createdAt
    PromptTemplateInput:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
        agentId:
          type: string
        category:
          type: string
      required:
        - name
        - content
        - category

    # ─── Files ───────────────────────────────────────────────────────────────
...