# lib/api-zod/src/generated/types/promptTemplate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HyperMind AI API — multi-agent AI SaaS platform
 * OpenAPI spec version: 0.1.0
 */

export interface PromptTemplate {
  id: number;
  name: string;
  content: string;
  /** @nullable */
  agentId?: string | null;
  category: string;
  isBuiltin: boolean;
  createdAt: Date;
}